+++ src/Language/Hasmtlib/Example/Bitvector.hs
@@ -2,23 +2,17 @@ module Language.Hasmtlib.Example.Bitvector where
 
 import Language.Hasmtlib
 import Data.Default
+import Data.Bits
 
 main :: IO ()
 main = do
   res <- solveWith @SMT (debug bitwuzla def) $ do
     setLogic "QF_BV"
 
-    xbv8 <- variable
-    ybv8 <- var @(BvSort 8)
+    x <- var @(BvSort 8)
 
-    assert $ true === (xbv8 `xor` ybv8)
-    assert $ xbv8 <=? maxBound
+    assert $ x === clearBit (maxBound `div` 2) 2
 
-    assert $ xbv8 >? 0
-    assert $ ybv8 >? 0
-
-    assert $ xbv8 + ybv8 >? xbv8 * ybv8
-
-    return (xbv8, ybv8)
+    return x
 
   print res